+++ src/PythonClient/src/quixstreams/models/__init__.py
@@ -1,5 +1,6 @@
 from .autooffsetreset import AutoOffsetReset
 from .codectype import CodecType
+from .codecsettings import CodecSettings
 from .commitmode import CommitMode
 from .commitoptions import CommitOptions
 from .eventdata import EventData
